+++ Fotogroep Waalre/View/PhotoClubsView/PhotoClubView.swift
@@ -139,11 +139,10 @@ struct PhotoClubView: View {
                     try viewContext.save()
                 }
             } catch {
-                // Replace this implementation with code to handle the error appropriately.
-                // fatalError() causes the application to generate a crash log and terminate.
-                // You should not use this function in a shipping application
                 let nsError = error as NSError
-                fatalError("Unresolved error \(nsError), \(nsError.userInfo)")
+                ifDebugFatalError("Unresolved error \(nsError), \(nsError.userInfo)",
+                                  file: #fileID, line: #line) // likely deprecation of #fileID in Swift 6.0
+                // in release mode, the failed deletion is only logged. App doesn't stop.
             }
         }