+++ Fotogroep Waalre/View/WhoIsWhoView/WhoIsWhoInnerView.swift
@@ -126,11 +126,10 @@ struct WhoIsWhoInnerView: View {
                 print("Deleted photographer \(fullName) and any associated memberships")
             }
         } catch {
-            // Replace this implementation with code to handle the error appropriately.
-            // fatalError() causes the application to generate a crash log and terminate.
-            // You should not use this function in a shipping application, although it may be useful during development.
             let nsError = error as NSError
-            fatalError("Unresolved error deleting photographer \(fullName): \(nsError), \(nsError.userInfo)")
+            ifDebugFatalError("Unresolved error deleting photographer \(fullName): \(nsError), \(nsError.userInfo)",
+                              file: #fileID, line: #line) // likely deprecation of #fileID in Swift 6.0
+            // in release mode, the failed deletion is only logged. App doesn't stop.
         }
     }